decompress: Use correct SIMD on Intel GPUs
Without this change, the decompression layer generates GPU hangs on
Intel Xe2 GPUs on Linux.

The layer creates specialized shaders for 8-, 16-, 32-, and 64-wide
SIMD. The "native" size for GDeflate is 32. For 32 and 64, the shaders
use subgroup operations like subgroupBallot. For 8 and 16, a different
mechanism is used. This mechanism uses shared memory and barriers
because it has to synchronize across multiple subgroups.

During pipeline creation, the layer uses subgroup size controls to
determine the minimum subgroup size supported by the implementation. It
the uses that to select the shader specialization. However, when it
creates the pipeline, it doesn't pass this requirement along. It lets
the implementation pick. Before Lunar Lake, Intel GPUs support 8, 16,
and 32, so the subgroupSize = 8 specialization would be selected. Given
that the decompression shader is quite large, it seems very likely that
register pressure would prevent the SIMD32 shader from being selected.

On Lunar Lake, sizes 16 and 32 are supported. The register file on Lunar
Lake is twice as large, and the decompression shader can easily fit in
the register space of SIMD32. The result is that the version of the
shader intended for sizes <= 16 is used with 32. When that happens, it
appears that the decompression shader gets stuck in an infinite loop,
and it is terminated by the kernel.

The fix has two parts. First, use a somewhat better heuristic for
selecting the subgroup size. On Intel Xe and earlier GPUs, 16 is
(likely) ideal. On Intel Xe2 GPUs, 32 is ideal. After selecting the
subgroup size, use `requiredSubgroupSize` to ensure the subgroup size
selected by the driver matches the subgroup size selected by the layer.

Ref: https://gitlab.freedesktop.org/mesa/mesa/-/issues/12743